move annotation creation code to (relevant) VideoAnnotation component
